Rework product layout
