+++ backend/forecastbox/products/__init__.py
@@ -9,6 +9,6 @@
 
 from .definitions import DESCRIPTIONS, LABELS
 
-from . import ensemble, forecast, thermal, simple
+from . import ensemble, forecast, standard, thermal, plots
 
-__all__ = ["DESCRIPTIONS", "LABELS", "ensemble", "forecast", "thermal", "extreme", "simple"]
+__all__ = ["DESCRIPTIONS", "LABELS", "ensemble", "forecast", "thermal", "extreme", "standard", "plots"]
