+++ backend/forecastbox/products/plots/__init__.py
@@ -1,0 +1,19 @@
+# (C) Copyright 2024- ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+
+from forecastbox.products.registry import CategoryRegistry
+from forecastbox.products.interfaces import Interfaces
+
+
+plot_product_registry = CategoryRegistry(
+    "Plots", interface=[Interfaces.STANDARD, Interfaces.DETAILED], description="Display products as plots", title="Plots"
+)
+
+from . import maps, meteogram  # noqa: F401, E402
