+++ backend/forecastbox/products/standard.py
@@ -7,10 +7,64 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
+
 from forecastbox.products.registry import CategoryRegistry
 
-from .interfaces import Interfaces
+from forecastbox.products.product import GenericTemporalProduct
+
+from forecastbox.models import Model
+from forecastbox.products.interfaces import Interfaces
+
 
 standard_product_registry = CategoryRegistry(
     "Standard", interface=[Interfaces.STANDARD, Interfaces.DETAILED], description="Standard products", title="Standard Products"
 )
+
+OUTPUT_TYPES = ["grib", "xarray"]
+
+
+@standard_product_registry("Output")
+class GribProduct(GenericTemporalProduct):
+    multiselect = {
+        "param": True,
+        "step": True,
+    }
+
+    @property
+    def qube(self):
+        return self.make_generic_qube(format=OUTPUT_TYPES)
+
+    def to_graph(self, product_spec, model, source):
+        source = self.select_on_specification(product_spec, source).concatenate("param").concatenate("step")
+        return source.map(self.named_payload("grib"))
+
+
+@standard_product_registry("Deaccumulated")
+class DeaccumulatedProduct(GenericTemporalProduct):
+    """
+    Deaccumulated Product.
+    """
+
+    multiselect = {
+        "param": True,
+        "step": True,
+    }
+
+    @property
+    def qube(self):
+        return self.make_generic_qube()
+
+    def model_intersection(self, model: Model):
+        """
+        Model intersection with the product qube.
+
+        Only the accumulation variables are used to create the intersection.
+        """
+        self_qube = self.make_generic_qube(param=model.accumulations)
+
+        intersection = model.qube(self.model_assumptions) & self_qube
+        result = f"step={'/'.join(map(str, model.timesteps))}" / intersection
+        return result
+
+    def to_graph(self, product_spec, model, source):
+        return self.select_on_specification(product_spec, model.deaccumulate(source)).map(self.named_payload("deaccumulated"))
